code cleanup, added file for columnHelper functions. '
